fix: upload img

+++ server/src/app/controllers/BlogController.js
@@ -16,9 +16,12 @@ const BlogController = {
     }
   },
   createBlog: async (req, res, next) => {
+    console.log("🚀 ~ createBlog: ~ req:", req);
     try {
       const { title, content, author } = req.body;
-      const attachment = req.file ? `/uploads/${req.file.filename}` : null;
+      const attachment = req.file
+        ? `/uploads/${req.file.filename}`
+        : "/uploads/default.png";
       const newBlog = new Blog({ title, content, author, attachment });
       await newBlog.save();
       res.status(200).json({
@@ -30,10 +33,21 @@ const BlogController = {
     }
   },
   updateBlog: async (req, res, next) => {
+    console.log("🚀 ~ updateBlog: ~ req:", req);
     try {
       const { id } = req.params;
       const { title, content, author } = req.body;
-      const attachment = req.file ? `/uploads/${req.file.filename}` : null;
+
+      const blog = await Blog.findById(id);
+      console.log("🚀 ~ updateBlog: ~ blog:", blog);
+
+      if (!blog) {
+        return res.status(404).json({ error: "Blog not found." });
+      }
+
+      const attachment = req.file
+        ? `/uploads/${req.file.filename}`
+        : blog.attachment;
 
       if (!title || !content) {
         return res
@@ -41,17 +55,20 @@ const BlogController = {
           .json({ error: "Title and content are required." });
       }
 
-      const blog = await Blog.findByIdAndUpdate(
+      if (req.file && !blog.attachment.endsWith("default.png")) {
+        const filePath = path.join(__dirname, "../../..", blog.attachment); // Đường dẫn file
+        if (fs.existsSync(filePath)) {
+          fs.unlinkSync(filePath); // Xóa file ảnh cũ
+          console.log(`Đã xóa file: ${filePath}`);
+        }
+      }
+      const updatedBlog = await Blog.findByIdAndUpdate(
         id,
         { $set: { title, content, author, attachment } },
-        { returnDocument: "after" }
+        { new: true }
       );
 
-      if (!blog) {
-        return res.status(404).json({ error: "Blog not found." });
-      }
-
-      res.status(200).json(blog);
+      res.status(200).json(updatedBlog);
     } catch (error) {
       console.error("Error updating blog:", error);
       res
@@ -74,7 +91,8 @@ const BlogController = {
           "../../..",
           deletedBlog.attachment
         ); // Đường dẫn file
-        if (fs.existsSync(filePath)) {
+
+        if (fs.existsSync(filePath) && !filePath.endsWith("default.png")) {
           fs.unlinkSync(filePath); // Xóa file ảnh
         }
       }
